+++ app/[locale]/(tabs)/requests/approve/[id]/page.tsx
@@ -1,13 +1,6 @@
 
-import { getAsset, getRequest, loadAllRequests } from "../../../../components/chain_utils";
+import { getAsset, getRequest } from "../../../../../../components/chain_utils";
 import ApprovePanel from "./panel";
-
-export async function generateStaticParams() {
-    const requests = await loadAllRequests();
-    return requests.map(request => ({
-        id: request.id,
-    }));
-}
 
 export default async function Page({ params }) {
     const request = await getRequest(params.id);