+++ app/[locale]/head.tsx
@@ -1,5 +1,4 @@
 'use client';
-import {useAppContext} from '../components/context';
 
 const i18n = {
   en: {
@@ -10,8 +9,18 @@ const i18n = {
   }
 }
 
-export default function Header() {
-  const { lang } = useAppContext();
+export default function Header({
+  params,
+}: {
+  params: {
+      locale: string;
+  }
+}) {
+  const { locale } = params;
+  let lang = 'cn';
+  if ('en-us' === locale){
+    lang = 'en';
+  }
   const texts = i18n[lang];
   return (
     <>